renamed substructure cfi to cff
